requirements.txt
Flask==2.1.0
kornia==0.6.5
MedPy==0.4.0
monai==0.8.1
nibabel==3.2.1
numpy==1.20.3
plotext==4.2.0
pytorch_lightning==1.6.3
setuptools==59.5.0
torchio==0.18.47
